control-room/backend/pyproject.toml
[project]
name = "control-room-backend"
version = "0.1.0"
description = "FastAPI service for Rozet control room"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.114.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.6.0",
  "sqlalchemy[asyncio]>=2.0.29",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.1",
  "python-jose[cryptography]>=3.3.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "orjson>=3.9.10"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.3",
  "black>=24.4.0"
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
